refactor/atualizando dtos para a inicialização do processo de compra

+++ super-mercado/src/main/java/org/example/DTO/jornada/JornadaDTO.java
@@ -3,27 +3,28 @@ package org.example.DTO.jornada;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class JornadaDTO {
     private LocalDateTime horarioAbertura;
     private LocalDateTime horarioFechamento;
-    private Double valorInicial;
-    private Double valorFinal;
+    private BigDecimal valorInicial;
+    private BigDecimal valorFinal;
     private Integer caixa;
     private String cpfFuncionario;
 
     public JornadaDTO() {
     }
 
-    public JornadaDTO(LocalDateTime horarioAbertura, Double valorInicial, Integer caixa, String cpfFuncionario) {
+    public JornadaDTO(LocalDateTime horarioAbertura, BigDecimal valorInicial, Integer caixa, String cpfFuncionario) {
         this.horarioAbertura = horarioAbertura;
         this.valorInicial = valorInicial;
         this.caixa = caixa;
         this.cpfFuncionario = cpfFuncionario;
     }
 
-    public JornadaDTO(LocalDateTime horarioFechamento, Double valorFinal, Integer caixa) {
+    public JornadaDTO(LocalDateTime horarioFechamento, BigDecimal valorFinal, Integer caixa) {
         this.horarioFechamento = horarioFechamento;
         this.valorFinal = valorFinal;
         this.caixa = caixa;
@@ -45,19 +46,19 @@ public class JornadaDTO {
         this.horarioFechamento = horarioFechamento;
     }
 
-    public Double getValorInicial() {
+    public BigDecimal getValorInicial() {
         return valorInicial;
     }
 
-    public void setValorInicial(Double valorInicial) {
+    public void setValorInicial(BigDecimal valorInicial) {
         this.valorInicial = valorInicial;
     }
 
-    public Double getValorFinal() {
+    public BigDecimal getValorFinal() {
         return valorFinal;
     }
 
-    public void setValorFinal(Double valorFinal) {
+    public void setValorFinal(BigDecimal valorFinal) {
         this.valorFinal = valorFinal;
     }
